requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-smart-selects==1.5.9
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
